docker: Creating a chroot Environment
With chroot you can set up and run programs or interactive shells such
as Bash in an encapsulated filesystem that is prevented from interacting
with your regular filesystem. Everything within the chroot environment
is penned in and contained. Nothing in the chroot environment can see
out past its own, special, root directory without escalating to root
privileges. That has earned this type of environment the nickname of a
chroot jail.

A chroot environment provides functionality similar to that of a virtual
machine, but it is a lighter solution.

Signed-off-by: Bill Liu <nonprivatemail@163.com>